refactor(engine): cleanup hook tests
